# lc-ssi-wc/openapi/swift-data-service.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Baseline SWIFT Data Service",
    "version": "1.1.0-prototype",
    "description": "SR2026 prototype contract. BIC syntax follows ISO 9362; currency codes are supplied by the ISO 4217 Currency service. Licensed SwiftRef/RMA/SSI data remains a production prerequisite."
  },
  "servers": [
    {
      "url": "http://localhost:3100/api"
    }
  ],
...
  "paths": {
    "/rma-authorisations": {
      "get": {
        "operationId": "listRma",
        "responses": {
          "200": {
            "description": "RMA records"
          }
        }
      },
      "post": {
        "operationId": "createRmaDraft",
        "requestBody": {
          "$ref": "#/components/requestBodies/Rma"
        },
        "responses": {
          "201": {
            "description": "DRAFT created"
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/rma-authorisations/{id}": {
      "put": {
        "operationId": "updateRmaDraft",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Rma"
        },
        "responses": {
          "200": {
            "description": "DRAFT updated"
          }
        }
      },
      "delete": {
        "operationId": "revokeRma",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "description": "Logically revoked; audit retained"
          }
        }
      }
    },
...
    "/ssis": {
      "get": {
        "operationId": "listSsi",
        "responses": {
          "200": {
            "description": "SSI records"
          }
        }
      },
      "post": {
        "operationId": "createSsiDraft",
        "requestBody": {
          "$ref": "#/components/requestBodies/Ssi"
        },
        "responses": {
          "201": {
            "description": "DRAFT created"
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/ssis/{id}": {
      "put": {
        "operationId": "updateSsiDraft",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Ssi"
        },
        "responses": {
          "200": {
            "description": "DRAFT updated"
          }
        }
      },
      "delete": {
        "operationId": "revokeSsi",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "description": "Logically revoked; audit retained"
          }
        }
      }
    },
...
    "/reference/fin-tag-suggestions": {
      "post": {
        "operationId": "suggestReferenceFin5xTags",
        "description": "Deprecated compatibility adapter. Use POST /reference/fin-tag-resolutions. It maps resolvedFields to the legacy suggestions shape and never supports payment execution.",
        "responses": {
          "200": {
            "description": "Non-payment MT 5x suggestions; paymentExecutable is always false."
          },
          "400": {
            "description": "FIN context incomplete or no verified exact mapping."
          }
        },
        "deprecated": true
      }
    },
...
    "/reference/fin-tag-resolutions": {
      "post": {
        "operationId": "resolveReferenceFin5xFields",
        "description": "Primary SR2026-governed, reference-only FIN field resolution for Treasury MT3 and Trade Finance MT4/MT7. No payment execution or confirmation semantics.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "service",
                  "resolutionMode",
                  "standardsRelease",
                  "messageType",
                  "direction",
                  "businessFunction",
                  "transactionReference",
                  "currency",
                  "receiverBic",
                  "valueDate",
                  "roles"
                ],
                "properties": {
                  "service": {
                    "const": "FIN"
                  },
                  "resolutionMode": {
                    "enum": ["TREASURY", "TRADE_FINANCE"]
                  },
                  "standardsRelease": {
                    "const": "SR2026"
                  },
                  "messageType": {
                    "type": "string",
                    "pattern": "^MT[0-9]{3}$"
                  },
                  "direction": {
                    "const": "OUTGOING"
                  },
                  "businessFunction": {
                    "type": "string"
                  },
                  "sequence": {
                    "type": "string"
                  },
                  "settlementLeg": {
                    "type": "string"
                  },
                  "fieldOptions": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  },
                  "transactionReference": {
                    "type": "string"
                  },
                  "currency": {
                    "type": "string",
                    "pattern": "^[A-Z]{3}$"
                  },
                  "receiverBic": {
                    "type": "string"
                  },
                  "bookingEntity": {
                    "type": "string"
                  },
                  "valueDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "roles": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  },
                  "roleEvidence": {
                    "type": "object",
                    "description": "Evidence and canonical settlement-route node identity for each role. Equal BICs are normalized only when canonicalRouteNodeId and ownerSide also match.",
                    "additionalProperties": {
                      "type": "object",
                      "properties": {
                        "ownerSide": {
                          "enum": [
                            "SENDER_SIDE",
                            "RECEIVER_SIDE",
                            "TRANSACTION_PARTY"
                          ]
                        },
                        "sourceType": { "type": "string" },
                        "sourceRecordId": { "type": "string" },
                        "version": { "type": "string" },
                        "status": { "type": "string" },
                        "approvalStatus": { "type": "string" },
                        "effectiveFrom": { "type": "string", "format": "date" },
                        "effectiveTo": { "type": "string", "format": "date" },
                        "canonicalRouteNodeId": {
                          "type": "string",
                          "description": "Identity derived from normalized BIC11, branch, account owner/servicer/identifier, currency, clearing context and instruction provenance; a BIC alone is insufficient."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reference-only resolvedFields result; paymentExecutable is always false.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "usage",
                    "paymentExecutable",
                    "confirmationSupported",
                    "resolvedFields"
                  ],
                  "properties": {
                    "usage": {
                      "const": "REFERENCE_ONLY"
                    },
                    "paymentExecutable": {
                      "const": false
                    },
                    "confirmationSupported": {
                      "const": false
                    },
                    "resolvedFields": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "standardsRelease",
                          "messageType",
                          "sequence",
                          "settlementLeg",
                          "tag",
                          "option",
                          "officialFieldName",
                          "officialRole",
                          "businessFunction",
                          "scopeStatus",
                          "resolutionStatus",
                          "reasonCode",
                          "resolvedValue",
                          "nvrRefs",
                          "provenance"
                        ],
                        "properties": {
                          "standardsRelease": {
                            "type": "string"
                          },
                          "messageType": {
                            "type": "string"
                          },
                          "sequence": {
                            "type": "string"
                          },
                          "settlementLeg": {
                            "type": "string"
                          },
                          "tag": {
                            "type": "string"
                          },
                          "option": {
                            "type": "string"
                          },
                          "officialFieldName": {
                            "type": "string"
                          },
                          "officialRole": {
                            "type": "string"
                          },
                          "businessFunction": {
                            "type": "string"
                          },
                          "scopeStatus": {
                            "enum": ["SSI_SUPPORTED", "OUT_OF_SSI_SCOPE"]
                          },
                          "resolutionStatus": {
                            "description": "Runtime disposition for every audited field slot. Profile-excluded slots are returned explicitly as N_A rather than omitted.",
                            "enum": [
                              "RESOLVED",
                              "NOT_REQUIRED",
                              "NO_ELIGIBLE_SSI",
                              "N_A"
                            ]
                          },
                          "reasonCode": {
                            "description": "Controlled reason such as ROUTE_COMPLETE or MESSAGE_PROFILE_EXCLUDED. A ROUTE_COMPLETE decision is attributed to ROUTE_RESOLVER/CANONICAL_ROUTE in provenance.",
                            "type": "string"
                          },
                          "resolvedValue": {
                            "type": ["string", "null"]
                          },
                          "nvrRefs": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "provenance": {
                            "description": "For NOT_REQUIRED route decisions, source is ROUTE_RESOLVER and ownerSide is CANONICAL_ROUTE; original candidate source/owner are retained separately when applicable.",
                            "type": "object"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Context, exact profile, option, or eligible SSI evidence is invalid/missing."
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "requestBodies": {
      "Rma": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RmaCommand"
            }
          }
        }
      },
      "Ssi": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SsiCommand"
            }
          }
        }
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "SWIFT/ISO syntax or business-rule validation failed"
      }
    },
    "schemas": {
...
      "Bic": {
        "type": "string",
        "pattern": "^[A-Z0-9]{4}[A-Z]{2}[A-Z0-9]{2}([A-Z0-9]{3})?$",
        "minLength": 8,
        "maxLength": 11,
        "description": "ISO 9362 BIC syntax only; directory existence requires licensed Bank/SwiftRef data."
      },
      "Currency": {
        "type": "string",
        "pattern": "^[A-Z]{3}$",
        "description": "Must also exist in Currency service ISO 4217 data."
      },
...
      "SwiftMessageType": {
        "type": "string",
        "pattern": "^(MT[0-9]{3}(COV)?|pacs\\.[A-Za-z0-9.]+|\\*)$"
      },
      "RmaCommand": {
        "type": "object",
        "required": [
          "ownBic",
          "counterpartyBic",
          "service",
          "direction",
          "messageTypes",
          "validFrom",
          "validTo",
          "maker"
        ],
        "properties": {
          "ownBic": {
            "$ref": "#/components/schemas/Bic"
          },
          "counterpartyBic": {
            "$ref": "#/components/schemas/Bic"
          },
          "service": {
            "enum": ["FIN", "FINPLUS"]
          },
          "direction": {
            "enum": ["INBOUND", "OUTBOUND"]
          },
          "messageTypes": {
            "type": "array",
            "minItems": 1,
            "uniqueItems": true,
            "items": {
              "$ref": "#/components/schemas/SwiftMessageType"
            }
          },
          "validFrom": {
            "type": "string",
            "format": "date"
          },
          "validTo": {
            "type": "string",
            "format": "date"
          },
          "maker": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "SsiCommand": {
        "type": "object",
        "required": ["counterpartyId", "scope", "route", "maker"],
        "properties": {
          "counterpartyId": {
            "type": "string",
            "minLength": 1
          },
          "scope": {
            "enum": ["STANDING", "TRANSACTION_SPECIFIC"]
          },
          "route": {
            "type": "object",
            "required": [
              "currency",
              "beneficiaryBic",
              "accountWithBic",
              "settlementCountry",
              "settlementMarket",
              "clearingSystem",
              "schemeType"
            ],
            "properties": {
              "currency": {
                "$ref": "#/components/schemas/Currency"
              },
              "beneficiaryBic": {
                "$ref": "#/components/schemas/Bic"
              },
              "accountWithBic": {
                "$ref": "#/components/schemas/Bic"
              },
              "intermediaryBic": {
                "$ref": "#/components/schemas/Bic"
              },
              "accountId": {
                "type": "string"
              },
              "settlementCountry": {
                "type": "string",
                "pattern": "^[A-Z]{2}$",
                "description": "ISO 3166-1 alpha-2 settlement-market country; distinct from counterparty/correspondent country."
              },
              "settlementMarket": {
                "type": "string",
                "minLength": 1
              },
              "clearingSystem": {
                "type": "string",
                "minLength": 1,
                "not": {
                  "enum": ["ANY"]
                },
                "description": "Concrete active standing-data code for the related RTGS or clearing scheme."
              },
              "schemeType": {
                "enum": ["RTGS", "LVPS", "ACH", "IPS"],
                "description": "Governed classification: RTGS, other large-value payment system, ACH, or instant payment scheme."
              }
            }
          },
          "maker": {
            "type": "string",
            "minLength": 1
          }
        }
      },
...
    }
  }
}